Refactor code and improve SQL schema for performance
This commit refines code and SQL schemas for better performance and readability. Notably, the 'mmsi' data type has been changed from BIGINT to INTEGER for performance improvement. In addition, the documentation has been updated and the readme file renamed for more clarity.

+++ aisdb/aisdb_sql/psql_createtable_dynamic_noindex.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS ais_{0}_dynamic (
-    mmsi BIGINT NOT NULL,
+    mmsi INTEGER NOT NULL,
     time INTEGER NOT NULL,
     longitude REAL NOT NULL,
     latitude REAL NOT NULL,
